Move VRP interfaces to other package
