+++ src/com/mjakobczyk/vrp/VrpDataProvider.java
@@ -1,4 +1,4 @@
-package com.mjakobczyk.vrp.def;
+package com.mjakobczyk.vrp;
 
 import com.mjakobczyk.vrp.model.VrpInput;
