+++ src/com/mjakobczyk/vrp/def/impl/DefaultVrpDataProvider.java
@@ -1,6 +1,6 @@
 package com.mjakobczyk.vrp.def.impl;
 
-import com.mjakobczyk.vrp.def.VrpDataProvider;
+import com.mjakobczyk.vrp.VrpDataProvider;
 import com.mjakobczyk.vrp.model.VrpInput;
 
 import java.util.Optional;
@@ -9,7 +9,7 @@ import java.util.logging.Logger;
 
 /**
  * DefaultVrpDataProvider is a default implementation of data provider
- * for solving Vehicle Routing Problem.
+ * for solving VRP.
  */
 public class DefaultVrpDataProvider implements VrpDataProvider {
