+++ src/com/mjakobczyk/vrp/def/impl/DefaultVrpSolutionProvider.java
@@ -1,6 +1,6 @@
 package com.mjakobczyk.vrp.def.impl;
 
-import com.mjakobczyk.vrp.def.VrpSolutionProvider;
+import com.mjakobczyk.vrp.VrpSolutionProvider;
 import com.mjakobczyk.vrp.model.VrpInput;
 import com.mjakobczyk.vrp.model.VrpOutput;
