+++ src/com/mjakobczyk/vrp/dynamic/impl/DynamicVrpSolver.java
@@ -1,8 +1,8 @@
 package com.mjakobczyk.vrp.dynamic.impl;
 
 import com.mjakobczyk.vrp.VrpSolver;
-import com.mjakobczyk.vrp.def.VrpDataProvider;
-import com.mjakobczyk.vrp.def.VrpSolutionProvider;
+import com.mjakobczyk.vrp.VrpDataProvider;
+import com.mjakobczyk.vrp.VrpSolutionProvider;
 import com.mjakobczyk.vrp.model.VrpOutput;
 
 import java.util.Optional;
